Requests: move the real functionality to a namespaced Requests class
Move the _real_ `Requests` class - with the exception of the autoload functionality - to the `src` directory and namespace it.

Includes updating the reference to this class in the release checklist.
