+++ .github/release-checklist.md
@@ -5,8 +5,8 @@ Template to use for release PRs from `develop` to `stable`
 
 PR for tracking changes for the x.x.x release. Target release date: **DOW MONTH DAY YEAR**.
 - [ ] Check if any dependencies need updating.
-- [ ] Check if the `cacert.pem` file needs updating.
-- [ ] Update the version constant in `library/Requests.php`.
+- [ ] Check if the `src/Transports/cacert.pem` file needs updating.
+- [ ] Update the version constant in `src/Requests.php`.
 - [ ] Add changelog for the release - PR #xxx
 - [ ] Merge this PR.
 - [ ] Make sure all CI builds are green.
